# Conflicts:
#	src/main/java/net/finmath/marketdata/model/curves/AbstractCurve.java
#	src/main/java/net/finmath/marketdata/model/curves/Curve.java
#	src/main/java/net/finmath/marketdata/model/curves/DiscountCurveNelsonSiegelSvensson.java
#	src/main/java/net/finmath/marketdata/model/curves/ForwardCurve.java
#	src/main/java/net/finmath/marketdata/model/curves/ForwardCurveFromDiscountCurve.java
#	src/main/java/net/finmath/marketdata/model/curves/ForwardCurveNelsonSiegelSvensson.java
#	src/main/java/net/finmath/marketdata/model/curves/SeasonalCurve.java
#	src/main/java/net/finmath/marketdata/model/volatilities/AbstractVolatilitySurface.java
#	src/main/java/net/finmath/marketdata/model/volatilities/CapletVolatilities.java
#	src/main/java/net/finmath/montecarlo/interestrate/modelplugins/AbstractLIBORCovarianceModel.java
#	src/main/java/net/finmath/montecarlo/interestrate/products/indices/AccruedInterest.java
#	src/main/java/net/finmath/montecarlo/interestrate/products/indices/DateIndex.java
#	src/main/java/net/finmath/montecarlo/interestrate/products/indices/TimeDiscreteEndOfMonthIndex.java
#	src/main/java/net/finmath/time/daycount/DayCountConventionFactory.java
#	src/main/java/net/finmath/time/daycount/DayCountConventionInterface.java
#	src/main/java/net/finmath/time/daycount/DayCountConvention_ACT_ACT_YEARFRAC.java
#	src/main/java6/net/finmath/marketdata/model/curves/AbstractCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/AbstractForwardCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/Curve.java
#	src/main/java6/net/finmath/marketdata/model/curves/CurveFromProductOfCurves.java
#	src/main/java6/net/finmath/marketdata/model/curves/CurveInterface.java
#	src/main/java6/net/finmath/marketdata/model/curves/DiscountCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/DiscountCurveFromProductOfCurves.java
#	src/main/java6/net/finmath/marketdata/model/curves/DiscountCurveNelsonSiegelSvensson.java
#	src/main/java6/net/finmath/marketdata/model/curves/ForwardCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/ForwardCurveFromDiscountCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/ForwardCurveNelsonSiegelSvensson.java
#	src/main/java6/net/finmath/marketdata/model/curves/PiecewiseCurve.java
#	src/main/java6/net/finmath/marketdata/model/curves/SeasonalCurve.java
#	src/main/java6/net/finmath/marketdata/model/volatilities/AbstractVolatilitySurface.java
#	src/main/java6/net/finmath/marketdata/model/volatilities/CapletVolatilities.java
#	src/main/java6/net/finmath/marketdata/model/volatilities/CapletVolatilitiesParametric.java
#	src/main/java6/net/finmath/marketdata/model/volatilities/CapletVolatilitiesParametricFourParameterPicewiseConstant.java
#	src/main/java6/net/finmath/marketdata/model/volatilities/VolatilitySurfaceInterface.java
#	src/main/java6/net/finmath/montecarlo/interestrate/modelplugins/AbstractLIBORCovarianceModel.java
#	src/main/java6/net/finmath/montecarlo/interestrate/products/indices/AccruedInterest.java
#	src/main/java6/net/finmath/montecarlo/interestrate/products/indices/DateIndex.java
#	src/main/java6/net/finmath/montecarlo/interestrate/products/indices/TimeDiscreteEndOfMonthIndex.java
#	src/main/java6/net/finmath/time/Period.java
#	src/main/java6/net/finmath/time/Schedule.java
#	src/main/java6/net/finmath/time/ScheduleGenerator.java
#	src/main/java6/net/finmath/time/ScheduleInterface.java
#	src/main/java6/net/finmath/time/Tenor.java
#	src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendar.java
#	src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendarAny.java
#	src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendarExcludingTARGETHolidays.java
#	src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendarExcludingWeekends.java
#	src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendarInterface.java
#	src/main/java6/net/finmath/time/daycount/DayCountConventionFactory.java
#	src/main/java6/net/finmath/time/daycount/DayCountConventionInterface.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_30E_360.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_30E_360_ISDA.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_30U_360.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_360.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_365.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_365A.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_365L.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_ACT_AFB.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_ACT_ICMA.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_ACT_ISDA.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_ACT_YEARFRAC.java
#	src/main/java6/net/finmath/time/daycount/DayCountConvention_NL_365.java
#	src/test/java/net/finmath/marketdata/model/curves/DiscountCurveNelsonSiegelSvenssonTest.java
#	src/test/java/net/finmath/marketdata/model/curves/NelsonSiegelSvenssonCalibrationTest.java
#	src/test/java/net/finmath/marketdata/model/volatilities/VolatilityCalibrationTest.java
#	src/test/java/net/finmath/tests/montecarlo/interestrate/CapValuationTest.java
#	src/test/java6/net/finmath/marketdata/model/curves/DiscountCurveNelsonSiegelSvenssonTest.java
#	src/test/java6/net/finmath/marketdata/model/curves/ForwardCurveNelsonSiegelSvenssonTest.java
#	src/test/java6/net/finmath/marketdata/model/curves/NelsonSiegelSvenssonCalibrationTest.java
#	src/test/java6/net/finmath/marketdata/model/volatilities/CapletVolatilitiesParametricTest.java
#	src/test/java6/net/finmath/marketdata/model/volatilities/VolatilityCalibrationTest.java
#	src/test/java6/net/finmath/montecarlo/interestrate/products/SwapLegTest.java
#	src/test/java6/net/finmath/montecarlo/interestrate/products/components/ExposureTest.java
#	src/test/java6/net/finmath/montecarlo/interestrate/products/indices/LIBORIndexTest.java
#	src/test/java6/net/finmath/tests/montecarlo/interestrate/CapValuationTest.java
#	src/test/java6/net/finmath/tests/time/daycount/DayCountConventionTest.java
#	src/test/java6/net/finmath/time/daycount/DayCountConvention_30E_360Test.java
#	src/test/java6/net/finmath/time/daycount/DayCountConvention_30E_360_ISDATest.java

+++ src/main/java6/net/finmath/time/daycount/DayCountConvention_30E_360_ISDA.java
@@ -44,19 +44,19 @@ public class DayCountConvention_30E_360_ISDA implements DayCountConventionInterf
 	}
 
 	/* (non-Javadoc)
-	 * @see net.finmath.time.daycount.DayCountConventionInterface#getDaycount(org.joda.time.LocalDate, org.joda.time.LocalDate)
+	 * @see net.finmath.time.daycount.DayCountConventionInterface#getDaycount(java.time.LocalDate, java.time.LocalDate)
 	 */
 	@Override
 	public double getDaycount(LocalDate startDate, LocalDate endDate) {
 		if(startDate.isAfter(endDate)) return -getDaycount(endDate,startDate);
 
-		double startDateDay 	= startDate.getDayOfMonth();
-		double startDateMonth 	= startDate.getMonthOfYear();
-		double startDateYear 	= startDate.getYear();
+		int startDateDay 	= startDate.getDayOfMonth();
+		int startDateMonth 	= startDate.getMonthOfYear();
+		int startDateYear 	= startDate.getYear();
 
-		double endDateDay 		= endDate.getDayOfMonth();
-		double endDateMonth 	= endDate.getMonthOfYear();
-		double endDateYear 		= endDate.getYear();
+		int endDateDay 		= endDate.getDayOfMonth();
+		int endDateMonth 	= endDate.getMonthOfYear();
+		int endDateYear 	= endDate.getYear();
 
 
 		// Check if we have last day of February
@@ -71,7 +71,7 @@ public class DayCountConvention_30E_360_ISDA implements DayCountConventionInterf
 	}
 
 	/* (non-Javadoc)
-	 * @see net.finmath.time.daycount.DayCountConventionInterface#getDaycountFraction(java.util.GregorianCalendar, java.util.GregorianCalendar)
+	 * @see net.finmath.time.daycount.DayCountConventionInterface#getDaycountFraction(java.time.LocalDate, java.time.LocalDate)
 	 */
 	@Override
 	public double getDaycountFraction(LocalDate startDate, LocalDate endDate) {
